^papirus

+++ scripts/papirus.sh
@@ -10,6 +10,8 @@ pkcon install papirus-icon-theme papirus-folders
 #sudo papirus-folders -C paleorange # XFCE
 
 # Tema Papirus para o LibreOffice
-wget -cq --show-progress https://launchpad.net/~papirus/+archive/ubuntu/papirus/+files/libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
-pkcon install-local ./libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
-rm ./libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
+if [ -f /usr/share/applications/libreoffice-startcenter.desktop ]; then
+    wget -cq --show-progress https://launchpad.net/~papirus/+archive/ubuntu/papirus/+files/libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
+    pkcon install-local ./libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
+    rm ./libreoffice-style-papirus_20180413-46+pkg3~ubuntu20.04.1_all.deb
+fi
